Make InteractiveStage handle more things
It now has its own AppearingContent component, because I'm always using
that. It also manages its buttons by itself now.
