+++ src/pages/stages/analysis.jsx
@@ -2,12 +2,9 @@
  * Poem analysis stages.
  */
 
-import React, { useState } from 'react'
+import React from 'react'
 
 import { InteractiveStage } from './stage'
-import AppearingContent from '../../util-components/appearingcontent'
-
-import { APS } from '../../constants'
 
 /** @typedef {import('./all').InteractionStageProps} InteractionStageProps */
 
@@ -15,17 +12,8 @@ import { APS } from '../../constants'
  * @param {InteractionStageProps} props
  */
 export function Perspective (props) {
-  const [showButtons, setShowButtons] = useState(false)
-
-  function onComplete () {
-    setTimeout(() => {
-      setShowButtons(true)
-    }, 2000)
-  }
-
   return (
     <InteractiveStage
-      showButtons={showButtons}
       nextStage={props.nextStage}
       increaseMistakes={props.increaseMistakes}
       buttons={[
@@ -33,27 +21,25 @@ export function Perspective (props) {
         { content: 'These lines reveal that the speaker is looking at the world.' },
         { content: 'These lines reveal that the speaker is African-American.', correct: true }
       ]}
-      key='is1'
+      key='ais1'
     >
-      <AppearingContent lettersPerSecond={APS} onComplete={onComplete} key='a3'>
-        You read the poem, looking for thematic elements you can discuss in whatever
-        Ms. Robinson throws at you next &mdash; probably a paragraph. You notice as well
-        that there's a few pieces of scrap paper in front of you, probably for your notes.
+      You read the poem, looking for thematic elements you can discuss in whatever
+      Ms. Robinson throws at you next &mdash; probably a paragraph. You notice as well
+      that there's a few pieces of scrap paper in front of you, probably for your notes.
 
-        <br />
-        <br />
+      <br />
+      <br />
 
-        On your first read-through, some lines immediately stand out to you.
+      On your first read-through, some lines immediately stand out to you.
 
-        <p className='poem-text'>
-          I look at the world<br />
-          From awakening eyes in a black face—<br />
-          ...<br />
-          I look then at the silly walls<br />
-          Through dark eyes in a dark face—<br />
-          ...<br />
-        </p>
-      </AppearingContent>
+      <p className='poem-text'>
+        I look at the world<br />
+        From awakening eyes in a black face—<br />
+        ...<br />
+        I look then at the silly walls<br />
+        Through dark eyes in a dark face—<br />
+        ...<br />
+      </p>
     </InteractiveStage>
   )
 }
@@ -62,17 +48,8 @@ export function Perspective (props) {
  * @param {InteractionStageProps} props
  */
 export function Segregation (props) {
-  const [showButtons, setShowButtons] = useState(false)
-
-  function onComplete () {
-    setTimeout(() => {
-      setShowButtons(true)
-    }, 2000)
-  }
-
   return (
     <InteractiveStage
-      showButtons={showButtons}
       nextStage={props.nextStage}
       increaseMistakes={props.increaseMistakes}
       buttons={[
@@ -80,19 +57,17 @@ export function Segregation (props) {
         { content: 'These lines reveal that the speaker sees a narrow space.' },
         { content: 'These lines reveal that the speaker faces segregation.', correct: true }
       ]}
-      key='is2'
+      key='ais2'
     >
-      <AppearingContent lettersPerSecond={APS} onComplete={onComplete} key='a4'>
-        After noting down the speaker's perspective, you jump back to the third,
-        fourth, and fifth lines.
-        <p className='poem-text'>
-          ...<br />
-          And this is what I see:<br />
-          This fenced-off narrow space<br />
-          Assigned to me.<br />
-          ...
-        </p>
-      </AppearingContent>
+      After noting down the speaker's perspective, you jump back to the third,
+      fourth, and fifth lines.
+      <p className='poem-text'>
+        ...<br />
+        And this is what I see:<br />
+        This fenced-off narrow space<br />
+        Assigned to me.<br />
+        ...
+      </p>
     </InteractiveStage>
   )
 }
@@ -101,17 +76,8 @@ export function Segregation (props) {
  * @param {InteractionStageProps} props
  */
 export function Metaphor (props) {
-  const [showButtons, setShowButtons] = useState(false)
-
-  function onComplete () {
-    setTimeout(() => {
-      setShowButtons(true)
-    }, 2000)
-  }
-
   return (
     <InteractiveStage
-      showButtons={showButtons}
       nextStage={props.nextStage}
       increaseMistakes={props.increaseMistakes}
       buttons={[
@@ -119,20 +85,18 @@ export function Metaphor (props) {
         { content: 'This has no name. I don\'t know what you mean.' },
         { content: 'This is a metaphor comparing oppression to walls.', correct: true }
       ]}
-      key='is3'
+      key='ais3'
     >
-      <AppearingContent lettersPerSecond={APS} onComplete={onComplete} key='a5'>
-        Then, you realize that the speaker mentions walls and oppression quite a
-        bit, and that this has a name.
-        <p className='poem-text'>
-          I look then at the silly walls<br />
-          ...<br />
-          And this is what I know:<br />
-          That all these walls oppression builds<br />
-          Will have to go!<br />
-          ...
-        </p>
-      </AppearingContent>
+      Then, you realize that the speaker mentions walls and oppression quite a
+      bit, and that this has a name.
+      <p className='poem-text'>
+        I look then at the silly walls<br />
+        ...<br />
+        And this is what I know:<br />
+        That all these walls oppression builds<br />
+        Will have to go!<br />
+        ...
+      </p>
     </InteractiveStage>
   )
 }
@@ -141,17 +105,8 @@ export function Metaphor (props) {
  * @param {InteractionStageProps} props
  */
 export function RealizeOppression (props) {
-  const [showButtons, setShowButtons] = useState(false)
-
-  function onComplete () {
-    setTimeout(() => {
-      setShowButtons(true)
-    }, 2000)
-  }
-
   return (
     <InteractiveStage
-      showButtons={showButtons}
       nextStage={props.nextStage}
       increaseMistakes={props.increaseMistakes}
       buttons={[
@@ -159,18 +114,16 @@ export function RealizeOppression (props) {
         { content: 'They suggest that the speaker can see now.' },
         { content: 'They suggest that the speaker is finally seeing the problems in their world.', correct: true }
       ]}
-      key='is3'
+      key='ais4'
     >
-      <AppearingContent lettersPerSecond={APS} onComplete={onComplete} key='a5'>
-        Next, you are drawn into these lines. Both talk of eyes.
-        <p className='poem-text'>
-          ...<br />
-          From awakening eyes in a black face—<br />
-          ...<br />
-          With eyes no longer blind—<br />
-          ...
-        </p>
-      </AppearingContent>
+      Next, you are drawn into these lines. Both talk of eyes.
+      <p className='poem-text'>
+        ...<br />
+        From awakening eyes in a black face—<br />
+        ...<br />
+        With eyes no longer blind—<br />
+        ...
+      </p>
     </InteractiveStage>
   )
 }
@@ -179,17 +132,8 @@ export function RealizeOppression (props) {
  * @param {InteractionStageProps} props
  */
 export function RealizeChange (props) {
-  const [showButtons, setShowButtons] = useState(false)
-
-  function onComplete () {
-    setTimeout(() => {
-      setShowButtons(true)
-    }, 2000)
-  }
-
   return (
     <InteractiveStage
-      showButtons={showButtons}
       nextStage={props.nextStage}
       increaseMistakes={props.increaseMistakes}
       buttons={[
@@ -197,21 +141,19 @@ export function RealizeChange (props) {
         { content: 'It reveals that the speaker can make their childhood dreams a reality.' },
         { content: 'It reveals that the speaker realizes change can happen.', correct: true }
       ]}
-      key='is4'
+      key='ais5'
     >
-      <AppearingContent lettersPerSecond={APS} onComplete={onComplete} key='a6'>
-        Finally, you look at the last stanza. You have a feeling that it's going to
-        be very, very powerful.
-        <p className='poem-text'>
-          ...<br />
-          I look at my own body<br />
-          With eyes no longer blind—<br />
-          And I see that my own hands can make<br />
-          The world that's in my mind.<br />
-          Then let us hurry, comrades,<br />
-          The road to find.
-        </p>
-      </AppearingContent>
+      Finally, you look at the last stanza. You have a feeling that it's going to
+      be very, very powerful.
+      <p className='poem-text'>
+        ...<br />
+        I look at my own body<br />
+        With eyes no longer blind—<br />
+        And I see that my own hands can make<br />
+        The world that's in my mind.<br />
+        Then let us hurry, comrades,<br />
+        The road to find.
+      </p>
     </InteractiveStage>
   )
 }
